requirements.txt
sentence-transformers
langchain
flask
pypdf
python-dotenv
pinecone-client  
langchain_community
langchain_google_genai
langchain_experimental
gunicorn
-e .
